Updated images and links. Also added a new statblock image for cluck
The image files were in the root folder for the site. Since the site
generates off the `content` folder, the images were not loading
correctly. I moved all the images to the misc/media folder.

 I also updated any links formatted as [[/folder/fileINeed]]. When
adding a link to a file just use the shorthand name for it in obsidian
and choose the selection from the dropdown that matches what you want.
Ideally no two files should be named the same but it happens sometimes.

I moved Lady Cluckingtons Skill-tree to its own file. I reference it in
clucks monster page.